requirements/requirement_production.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
click==8.1.3
Django==4.1.4
django-autoslug==1.9.8
django-countries==7.5
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
flake8==6.0.0
isort==5.11.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumbers==8.13.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
